+++ auth0_flutter/lib/src/mobile/web_authentication.dart
@@ -144,21 +144,27 @@ class WebAuthentication {
   /// versions of iOS and macOS. Requires an Associated Domain configured with
   /// the `webcredentials` service type, set to your Auth0 domain –or custom
   /// domain, if you have one.
-  Future<void> logout({
-    final String? returnTo,
-    final bool useHTTPS = false,
-    final bool federated = false,
-  }) async {
-    await Auth0FlutterWebAuthPlatform.instance.logout(
-      _createWebAuthRequest(
-        WebAuthLogoutOptions(
+  /// * (android only): [allowedBrowsers] Defines an allowlist of browser
+  /// packages
+  /// When the user's default browser is in the allowlist, it uses the default
+  /// browser
+  /// When the user's default browser is not in the allowlist, but the user has
+  /// another allowed browser installed, the allowed browser is used instead
+  /// When the user's default browser is not in the allowlist, and the user has
+  /// no other allowed browser installed, an error is returned
+  Future<void> logout(
+      {final String? returnTo,
+      final bool useHTTPS = false,
+      final List<String> allowedBrowsers = const [],
+      final bool federated = false}) async {
+    await Auth0FlutterWebAuthPlatform.instance.logout(_createWebAuthRequest(
+      WebAuthLogoutOptions(
           returnTo: returnTo,
           scheme: _scheme,
           useHTTPS: useHTTPS,
           federated: federated,
-        ),
-      ),
-    );
+          allowedBrowsers: allowedBrowsers),
+    ));
     await _credentialsManager?.clearCredentials();
   }
 
@@ -170,11 +176,12 @@ class WebAuthentication {
     Auth0FlutterWebAuthPlatform.instance.cancel();
   }
 
-  WebAuthRequest<TOptions> _createWebAuthRequest<
-    TOptions extends RequestOptions
-  >(final TOptions options) => WebAuthRequest<TOptions>(
-    account: _account,
-    options: options,
-    userAgent: _userAgent,
-  );
+  WebAuthRequest<TOptions>
+      _createWebAuthRequest<TOptions extends RequestOptions>(
+              final TOptions options) =>
+          WebAuthRequest<TOptions>(
+            account: _account,
+            options: options,
+            userAgent: _userAgent,
+          );
 }