+++ auth0_flutter/macos/Classes/AuthAPI/AuthAPICustomTokenExchangeMethodHandler.swift
@@ -1,0 +1,1 @@
+../../../darwin/Classes/AuthAPI/AuthAPICustomTokenExchangeMethodHandler.swift