+++ auth0_flutter_platform_interface/lib/src/web-auth/web_auth_logout_options.dart
@@ -5,18 +5,21 @@ class WebAuthLogoutOptions implements RequestOptions {
   final String? returnTo;
   final String? scheme;
   final bool federated;
+  final List<String> allowedBrowsers;
 
   WebAuthLogoutOptions(
       {this.useHTTPS = false,
       this.returnTo,
       this.scheme,
-      this.federated = false});
+      this.federated = false,
+      this.allowedBrowsers = const []});
 
   @override
   Map<String, dynamic> toMap() => {
         'useHTTPS': useHTTPS,
         'returnTo': returnTo,
         'scheme': scheme,
-        'federated': federated
+        'federated': federated,
+        'allowedBrowsers': allowedBrowsers
       };
 }